% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{beta_adjust_bin}
\alias{beta_adjust_bin}
\title{Adjusts the value of the binomial coefficients to account for the scaling of x.}
\usage{
beta_adjust_bin(beta, sigma_x)
}
\arguments{
\item{beta}{The estiamte of beta with scaled data.}

\item{sigma_x}{Sample tandard deviations of the original predictors.}
}
\value{
Returns the adjusted coefficients
}
\description{
Adjusts the value of the binomial coefficients to account for the scaling of x.
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
