% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{bin_horse}
\alias{bin_horse}
\title{The workhorse function for the binomial updates in mcen.  It uses IRWLS glmnet updates to solve the regression problem.}
\usage{
bin_horse(Y, X, delta, gamma_y, y_clusters, set_length, eps, maxiter)
}
\arguments{
\item{Y}{the matrix of responses}

\item{X}{the matrix of predictors with the intercept included}

\item{delta}{the tuning parameter for the lasso penalty}

\item{gamma_y}{the tuning parameter for the ridge fusion penalty}

\item{y_clusters}{the cluster assignments from the provided clustering algorithm}

\item{set_length}{the size of each cluster corresponding to a given response.  r dimensions with each element containing the cluster size of that responses cluster.}

\item{eps}{the tolerance for conversion normally 1e-5}

\item{maxiter}{the maximum number of iterations}
}
\value{
Returns a matrix of coefficients
}
\description{
The workhorse function for the binomial updates in mcen.  It uses IRWLS glmnet updates to solve the regression problem.
}
\author{
Brad Price <brad.price@mail.wvu.edu>
}
