% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{CalcHorseBin}
\alias{CalcHorseBin}
\title{Creates the the working response for all responses for glmnet binomial family}
\usage{
CalcHorseBin(Y, X, Beta)
}
\arguments{
\item{Y}{is the matrix of responses
result is the list of vectors needed for the working responses in glmnet}

\item{X}{the matrix of predictors.}

\item{Beta}{current iteration of the regression coefficients}
}
\description{
Creates the the working response for all responses for glmnet binomial family
}
\author{
Brad Price <brad.price@mail.wvu.edu>
}
