% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{CalcHorseEBin}
\alias{CalcHorseEBin}
\title{Creates the probabilities and working response for the glmnet update for a given response with a binomial family}
\usage{
CalcHorseEBin(X, Beta, Y, r)
}
\arguments{
\item{X}{the matrix of predictors.}

\item{Beta}{current iteration of the regression coefficients}

\item{Y}{is the matrix of responses}

\item{r}{the response of interest
result is a list of things needed for the working response in glmnet}
}
\description{
Creates the probabilities and working response for the glmnet update for a given response with a binomial family
}
\author{
Brad Price <brad.price@mail.wvu.edu>
}
