% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{cluster}
\alias{cluster}
\title{Wrapper function for different clustering methods}
\usage{
cluster(x, cNum, clusterMethod = "kmeans", clusterIterations = 100,
  clusterStartNum = 30)
}
\arguments{
\item{x}{data to be clustered. Clustering will be done on the columns.}

\item{cNum}{number of cluster centers}

\item{clusterMethod}{"kmean" for kmeans function, "kmeanspp" for kcca implementation of kmeans++}

\item{clusterIterations}{number of maximum iterations for clustering}

\item{clusterStartNum}{random number of starting points used}
}
\value{
Returns cluster assignments
}
\description{
Wrapper function for different clustering methods
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
