% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{matrix_multiply}
\alias{matrix_multiply}
\title{matrix multiply}
\usage{
matrix_multiply(beta, x)
}
\arguments{
\item{beta}{Matrix of coefficients.}

\item{x}{Design matrix.}
}
\value{
Returns x times beta
}
\description{
matrix multiply
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
