% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{mcen.init}
\alias{mcen.init}
\title{Provides initial estimates for the mcen functionF}
\usage{
mcen.init(x, y, family = "mgaussian", delta = NULL, gamma_y = 1,
  intercept = FALSE)
}
\arguments{
\item{x}{the n x p design matrix}

\item{y}{the n x y matrix of responses}

\item{family}{type of likelihood used two options "mgaussian" or "mbinomial"}

\item{delta}{sparsity tuning parameter}

\item{gamma_y}{tuning parameter for clustering responses}

\item{intercept}{whether an intercept should be included in the model}
}
\value{
matrix of coefficients
}
\description{
Provides initial estimates for the mcen functionF
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
