% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{pred_eval.mbinom_mcen}
\alias{pred_eval.mbinom_mcen}
\title{Evaluates prediction error for multiple binomial responses.}
\usage{
\method{pred_eval}{mbinom_mcen}(obj, test_x, test_y)
}
\arguments{
\item{obj}{The mbinom_mcen object.}

\item{test_x}{A matrix of the test predictors.}

\item{test_y}{A matrix of the test responses.}
}
\description{
Evaluates prediction error for multiple binomial responses.
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
