% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{pred_eval.mgauss_mcen}
\alias{pred_eval.mgauss_mcen}
\title{Calculates the prediction error for a mgauss_mcen object.}
\usage{
\method{pred_eval}{mgauss_mcen}(obj, test_x, test_y)
}
\arguments{
\item{obj}{The mgauss_mcen object.}

\item{test_x}{The matrix of test predictors.}

\item{test_y}{The matrix of test responses.}
}
\description{
Calculates the prediction error for a mgauss_mcen object.
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
