% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{pred_eval}
\alias{pred_eval}
\title{Calculates the out of sample likelihood for an mcen object}
\usage{
pred_eval(obj, test_x, test_y)
}
\arguments{
\item{obj}{The mcen object.}

\item{test_x}{The matrix of test predictors.}

\item{test_y}{The matrix of test responses.}
}
\description{
Calculates the out of sample likelihood for an mcen object
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
