% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{print.cv.mcen}
\alias{print.cv.mcen}
\title{Prints nice output for a cv.mcen object.}
\usage{
\method{print}{cv.mcen}(x, ...)
}
\arguments{
\item{x}{The cv.mcen object.}

\item{...}{Additional parameters.}
}
\value{
Prints out information about where the cv.mcen object was minimized.
}
\description{
Prints nice output for a cv.mcen object.
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
