% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{print.mcen}
\alias{print.mcen}
\title{Prints nice output for an mcen object.}
\usage{
\method{print}{mcen}(x, ...)
}
\arguments{
\item{x}{The mcen object.}

\item{...}{Additional parameters.}
}
\value{
Prints out some basic information about the mcen object.
}
\description{
Prints nice output for an mcen object.
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
