% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{randomly_assign}
\alias{randomly_assign}
\title{randomly assign n samples to k groups}
\usage{
randomly_assign(n, k)
}
\arguments{
\item{n}{number of samples}

\item{k}{number of groups}
}
\value{
Returns assignments of n into k groups
}
\description{
randomly assign n samples to k groups
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
