% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{SetEq}
\alias{SetEq}
\title{SetEq test set equivalence of two clustering sets}
\usage{
SetEq(set1, set2)
}
\arguments{
\item{set1}{is the cluster assignments of the previous iteration}

\item{set2}{is the cluster assignments of the current clusters}
}
\value{
Returns a logical saying if the two clusterings are equal
}
\description{
SetEq test set equivalence of two clustering sets
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
