% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{squared_error}
\alias{squared_error}
\title{Calculates sum of squared error between two vectors or matrices}
\usage{
squared_error(pred, test_y)
}
\arguments{
\item{pred}{the predictions}

\item{test_y}{the testing response values}
}
\value{
returns the sum of the squared differences between pred and test_y
}
\description{
Calculates sum of squared error between two vectors or matrices
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>, Brad Price <brad.price@mail.wvu.edu>
}
