% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcen.r
\name{vl_binom}
\alias{vl_binom}
\title{Calculates out of sample error on the binomial likelihood}
\usage{
vl_binom(pred, test_y)
}
\arguments{
\item{pred}{The predicted values.}

\item{test_y}{The test response values.}
}
\description{
Calculates out of sample error on the binomial likelihood
}
\author{
Brad Price <brad.price@mail.wvu.edu>
}
